SICD/AutoICD: forward types to avoid includes
move helper template to separate private header to avoid
use of symbols in header
